For the partially opaque case drop the boundary stroke for filled regions. This change makes example 30 look much better (although still not perfect) and leaves other examples that use filled regions (such as 8, 16, and 20) unaffected (i.e., they do have boundary strokes) since they are fully opaque fills.  I also tried the experiment of turning off the boundary stroking in the opaque case, but that made example 16 (and presumably 8 and 20 as well) look bad (with a background that leaked through due to AA) so that was only a temporary experiment.
I also played with the crispEdges shape-rendering attribute.  The previous
commented out version of this svg_attr_value function call had a syntax
error which I fixed, and I had to change crisp-edges to crispEdges in order
to make the result validate properly.  However, that attribute did not seem
to make any practical difference (at least for the svg viewers accessible to
me) so I commented it out again on the principle of using defaults if you
don't completely understand what is going on.  :-)


svn path=/trunk/; revision=9232