Preliminary fix for the test program under Windows (especially older versions of the MS Visual C/C++ compiler).
Applying the patch by Terrence was not enough:
- The MSVC 6.0 compiler does not recognise "long long int" as a valid type
  (So for all MSVC compilers, the type "__int64" is used instead)
- Running the qsastime_testlib program is unsuccessful with both bare
  Windows and Cygwin, as time_t is too small a data type (size = 4 bytes)
- The patch assumes that _putenv() exists, but the leading underscore
  is unusual for Linux/UNIX, where putenv() is the standard.

Tests to be added to fix this more properly: test for setenv(), putenv(),
long long int versus __int64. This means a lot of extra tests and then the
program will refuse to work anyway.

For the moment simply committing this patched version without CMake tests
to get the build going on Windows again.


svn path=/trunk/; revision=9556