Apply Tony Allen's patch to properly normalize strfMJD results using pre-rounding.  The implementation uses the most lightly rounded seconds value in the format string to set the degree of rounding used.  This change should yield proper normalization of the strfMJD formatted results in all cases.
To give an example, that is tested with Test 05 of qsastime_testlib,

if H:M:S = 0:59:59.99999999000, internally, that will format as a properly
normalized 1:00:00.0000 if sufficient rounding (e.g., %S%4 which corresponds
to rounding errors of 50 ms) is used, and will format as a properly
normalized 0:59:59.999999990 if light rounding (e.g., %S%9 which corresponds
to rounding errors of 0.5 ns) is used.  In the latter case, just like in any
floating-point calculation, reducing the rounding means you will see more
details, but those details potentially include numerical errors.


svn path=/trunk/; revision=9641