Implement page 4 of example to plot TAI-UTC as a function of time from 1950 to 2020. Initial results look like the basic leap-second code for plbtime ==> btimeqsas (including the table lookup using bhunt_search of TAI-UTC results within libqsastime) is working.
However, fine details near the UTC discontinuities undoubtedly need more
work for btimeqsas.  Also, leap seconds for ctimeqsas and strfqsas still
need to be implemented.  Also, the X-axis labelling needs a fix to produce
year labels on years which are multiples of 20 (i.e., 1960, 1980, 2000),
rather than shifted by 8 years from that (i.e, 1968, 1998,...) by default. I
expect it is a core issue with what the default is for the time labelling
case.


svn path=/trunk/; revision=9648