Add int if60secformat to strfMJD argument list.  This forces renormalization of results so that the integer part of the second is 60 to signal that the epoch is in the middle of leap-interval insertion just prior to a positive discontinuity in TAI-UTC.  The implementation of the actual renormalization is still in progress, but I am making the commit now because this builds, and all the other changes in this source file have been thoroughly tested and do work.
Return inleap value from leap_second_TAI to signal when the epoch is in the
middle of a leap-interval insertion just prior to a positive discontinuity
in TAI-UTC.  This inleap logic should be complete, but needs testing when
the if60secformat implementation is complete for strfMJD.

bhunt_search declaration moved to qsastimeP.h.

Simplify normalize_MJD argument list.

Drop "p" form of MJDtime variable name.  Instead, use _value prefix on value
form of variable and raw name for pointer to that value.


svn path=/trunk/; revision=9660