AWI for Alban Rochel.  Major update/reorganization.
Changes:
- Qt-based font support
- alpha blending
- fixed memory leaks, at least every one I could identify. Among the leaks
and problems identified by valgrind, some are related to Qt, but they appear
even on minimal Qt examples (a simple empty QApplication).
- hopefully fixed all crashed and segfaults
- fixed non-filled areas problems on example 30
- Large clean-up of the driver architecture. Working on the fonts, I've
found Qt classes allowing me to simplify *much* the interactive driver. The
QtPLBufferedDriver class is not used any more.
- the output size parameter is taken into account (I had forgotten this one
in the previous releases, letting everything to the default A4 size), except
for EPS/PDF: Qt 4.2 and 4.3 require a standard paper size. I've left it to
A4.

Bugs:

- I've had lots of issues with the SVG driver to have correct text
placement.  I've decided to give up for the moment, unless I've missed
something I believe I've found a bug in Qt. I've made a quick and dirty fix
which still leaves imperfect text offsets. I'll investigate a little, and do
a bug report to Qt if it is indeed a Qt bug. [AWI comment: our experience is
this symptom is often a result of text placement bugs in Linux SVG viewers.
For example, anything related to GNOME (eog) or ImageMagick uses the
librsvg2 library which has a bug with regard to SVG text placement.  For
reliable text-placement results use konqueror to view *.svg files.  When I
do that for example 2, I find the numbers displaced to the right a small
amount, but that may be due to Alban's "quick and dirty" fix to match bugs
in the SVG viewer he used.]

- No symbols on the x01 plots for the raster, EPS/PDF and interactive
drivers. No idea why, I remember seeing them at some point (!), but for the
moment (until the next release), let's forget them...

- The number of displayed symbols in example 23 is much smaller than with
other devices. I don't know much of font handling and unicode, so I don't
know if it is a Qt limitation or if I can do something about it (which I
suppose). Again, I'll deal with that in a future release.

- [AWI] Characters too small by roughly a factor of two.


svn path=/trunk/; revision=9729