AWI for Alban Rochel.  Alban got access to a Mandriva virtual machine which appeared to give results equivalent to AWI's Debian Lenny screenshots for the previous commit.
The present commit is his adjustments to make the Mandriva results look good
(and hopefully AWI's as well).

- Adjust glyph size to larger values.

- Qt gave bounding boxes a bit too large for the text, mainly on the Y axis,
and the text is not exactly centered vertically. Alban adjusted the bounding
box empirically to get better results. This should reduce the offset issues.

- The -geometry parameter is taken into account for the epsqt and qtwidget
devices.

(AWI's comments.  The changes improve a lot of things, but they make svgqt
(created with qt-4.5.0 and viewed with reliable versions of konqueror and
firefox) much worse.  Also, this commit fixed an issue with the epsqt
background, but the bounding box is quite bad now [almost two times too
large] for that case.  So I will play with these adjustments some more, but
for now it is good to get Alban's current best shot at an adjustment
committed to give us a benchmark for further effort.)


svn path=/trunk/; revision=9749