AWI for Alban Rochel.
Specify empty family name to setFamily so that setStyleHint is always
followed.

Drop empirical adjustments for the bounding box.

Reduce empirical size correction for fonts from 1.6 to 1.45.

Preliminary tests of all these changes with qtwidget show substantial
improvements.  For the first time, the glyph selection seems identical
to that of the cairo device.  The text offset problems also seem to be
gone.  A detailed report will follow on list.


svn path=/trunk/; revision=9797