I think there is bad logic in the IF(NOT CMAKE_MODULE_EXISTS) logic block since the relevant LIBRARY variables that are being copied to the equivalent MODULES variables are actually defined later.  PLplot doesn't use Ada modules so this bad logic is never exercised for that case.
There is no such logic block in the Fortran case or the D case so I have
removed the whole IF(NOT CMAKE_MODULE_EXISTS) logic block for the Ada case.
This should give better CMake support for Ada in general. As expected (since
no modules are used), the Ada bindings and examples builds continue to work
fine on Linux after this change.


svn path=/trunk/; revision=9909