Go back to Andrew's general fix for the possibility of debug or optimized keywords in library lists being exported to *.pc files since that is a better solution than mine which made some incorrect assumptions which could lead to problems on non-Linux platforms for -DCMAKE_BUILD_TYPE=Release.
I have checked that with this fix, the configured plplotd-qt.pc files are
identical for no specified CMAKE_BUILD_TYPE, -DCMAKE_BUILD_TYPE=Debug, and
-DCMAKE_BUILD_TYPE=Release for my particular Qt4.5.1 installation where
there is no real difference between the debug and optimized versions of the
Qt4 libraries (as shown by the printed out QT_LIBRARIES variable). This
result confirms the logic is working properly.  Of course, for other Qt4
installations with real differences between debug and optimized, the
configured plplotd-qt.pc file will depend on whether CMAKE_BUILD_TYPE is
Debug or not.


svn path=/trunk/; revision=9967