Split former qt.cpp into device driver part (qt.cpp) and library part plqt.cpp).  Much thanks to Dmitri Gribenko for the patch splitting the code this way.
Add CMake infrastructure to support this source-code split between the qt
device driver and libplplotqt.  

Running an example (30 in this case) for qtwidget, epsqt, pngqt, and svgqt
(the four kinds of qt devices) under valgrind showed 0 errors.  qt_example
was valgrind clean as well. N.B. The above tests were done for default CMake
configuration (ENABLE_DYNDRIVERS ON, ENABLE_qt ON) and with the new build
system for the installed examples and more testing for other configurations
should be done.


svn path=/trunk/; revision=10073