Work around CMake enable_language(<language> OPTIONAL) bug (where it errors out for missing or bad compilers rather than returning with CMAKE_<languageName>_COMPILER_WORKS as a status).  The idea of the workaround is to first search for the compiler.  If it does not exist then disable the appropriate component of PLplot with a WARNING message. Only if the compiler exists is there an attempt made to run the CMake enable_language language command to check the compiler.
N.B. This workaround only works if the compiler is completely missing.  Ada
is ideosyncratic so this workaround does not solve the issue in that case.
Also, this workaround does not solve the issue for existing but bad
compilers that give results that fail for the try_compile that is run by the
enable_language command.


svn path=/trunk/; revision=10096