Remove attempted workaround for CMake enable_language OPTIONAL signature bug which is described at http://public.kitware.com/Bug/view.php?id=9220.  Since that OPTIONAL signature was not working, the idea of the workaround was to pre-search for the compiler by using "include(CMakeDetermine<language>Compiler)".  If this check revealed the compiler was missing, we could then avoid the call to enable_language (which incorrectly errors out rather than warns when a compiler is missing).
However, it turns out this workaround causes cache issues (undefined
language-related variables) if the ccmake CLI or cmake-gui GUI are used
rather than the cmake application for configuring PLplot.  Thus, drop the
workaround and accept the hard landing (CMake error) instead.  This hard
landing will occur whenever a user fails to disable a language (Ada, C++, D,
Fortran, or Java) when the corresponding language development environment 
(compiler and associated libraries) is not installed on his computer.

The fundamental solution to this issue is to get the CMake developers to
address bug 9220 so the OPTIONAL signature of enable_language actually
avoids erroring out when the language development environment is
missing/broken. But this will take some time.


svn path=/trunk/; revision=10153