Big change to clean up the plplot namespace.
Now plConfig.h only contains macros which need to be externally visible because they are referenced in plplot.h or are used in the examples. These are all prefaced with PL_ .

All other macro definitions which are only used internally reside in config.h. 

Also, the HAVE_ISINF, HAVE_ISNAN and HAVE_FINITE definitions have PL_ added on the front to remove the redefinition warnings caused by python leaking the HAVE_* macros. 



svn path=/trunk/; revision=10169