Handle the install rpath for libplplotqt(d).so the same way it is done for qt.so.
The order is 

(1) installed PLplot location, 

(2) Qt location,

(3) Special rpath location for libplplot(d) which currently is the
location for libltdl (normally but not always the system location [/usr/lib
on Linux]).

If (1) is the system location,then qt and libltdl will also be taken from
the system location if they exist there.  However, this should usually not
be a problem since this case normally only occurs when all locations are
system locations.

If (2) is the system location, then libltdl will also be taken from
the system location if that exists.  This is probably not a major issue.

The correct way to solve this is to exclude system locations from rpath
locations, and presumably that can be done using the NO_CMAKE_SYSTEM_PATH
signature for find_library, but I haven't tried that since for now the
present RPATH order should be good enough for most purposes.


svn path=/trunk/; revision=10171