Drop ENABLE_cxx logic or replace it by CMAKE_CXX_COMPILER_WORKS logic where appropriate.
The result should be that cmake always tries to find a working C++ compiler
for a fresh build and continues to do that for subsequent "dirty" builds if
a working C++ compiler was not found previously. This logic sets ENABLE_cxx
OFF when there is no working C++ compiler, but the logic does not depend on
ENABLE_cxx.  Thus, that variable is strictly reserved for toggling whether
or not C++ bindings are to be built/installed and whether C++ examples are
built (for BUILD_TEST=ON) and installed.


svn path=/trunk/; revision=10201