Off-screen rendering for xcairo.  This provides a large speed increase.
A possibly negative side effect of this patch is that the plot is not displayed
in the plot window until either plflush is called or we reach the end of the
page.

A positive side effect of this patch is that xcairo no longer flickers
repeatedly due to replots on expose events (when the window has to be redrawn).


svn path=/trunk/; revision=10329