Protect all device driver calls using the (now) reentrant plsave_set_locale and plrestore_locale. The former saves the fundamental PLplot LC_NUMERIC locale and then sets the LC_NUMERIC locale to "C" for everything done by all device drivers.  Of course, any library called by a device driver can also change the locale, but we guard against such changes affecting the rest of PLplot by using plrestore_locale to restore the fundamental PLplot LC_NUMERIC locale saved by plsave_set_locale.
N.B. this logic allows the fundamental PLplot LC_NUMERIC locale (except for
the parts like the device drivers and palette file reading that are
protected by the combination of plsave_set_locale and plrestore_locale) to
be any valid locale set by any application or library that calls the PLplot
library.  If that locale specifies comma decimal separators rather than
period decimal separators, for example, then the PLplot core will
automatically use those for for specifying axis labels.  Those commas for
axis labels are drawn by the PLplot core for the Hershey font device drivers
or just propagate to the unicode device drivers as UCS4 arrays.  Thus, in
both cases, we get a comma decimal separator for axis labels (if that is
what the fundamental PLplot LC_NUMERIC locale calls for) independently of
the logic of the present commit that sets the LC_NUMERIC locale to "C" for
all device driver code that is executed.


svn path=/trunk/; revision=10383