Implement most of a common testing framework for the build-tree examples and the install-tree examples with full dependency tracking.  The current status is the following targets work:
... test_ada_psc
... test_all_diff_psc
... test_c_psc
... test_cxx_psc
... test_d_psc
... test_f77_psc
... test_f95_psc
... test_java_psc
... test_lua_psc
... test_ocaml_psc
... test_octave_psc
... test_pdl_psc
... test_python_psc
... test_some_diff_psc
... test_tcl_psc

test_all_diff_psc is an especially useful target for getting a summary
report (with all dependencies automatically satisfied) of how stdout and
PostScript results for our complete set of standard examples for all non-C
languages we have implemented compares with the corresponding C results.
This common testing framework is currently incomplete because there are
still known issues to be solved for the test_noninteractive and
test_interactive targets that are to be used for more
comprehensive testing than test_all_diff_psc.


svn path=/trunk/; revision=10435