Support systems without ocamlopt; Build examples with ocamlopt when available
PLplot's OCaml support will now build if ocamlopt is not present.  The
ocamlopt compiler is not available for as many OS and architecture
combinations as ocamlc.

The build-tree examples are now built with ocamlopt when it is available.
This produces slightly faster executables, though there isn't much of a
difference for most of the examples since most of the execution time for
the examples is spent in the C plotting functions.

Finally, the native code versions of the OCaml compilers, ocamlc.opt
and ocamlopt.opt, are used if they are available.  This provides a
significant speedup in compilation of the OCaml bindings and examples.


svn path=/trunk/; revision=10448