Change plplot_core ==> plplot in target names.
Do not combine file depends for the OCAMLOPT chain with those from the
OCAMLC chain when exporting the file depends associated with the
plplot_ocaml target.  Instead, rely just on file depends of either
target_plplot_cmxa (OCAMLOPT is true) or target_plplot_cma (OCAMLOPT is
false).  In other words, we assume the chain of file depends associated with
each of these targets is correct.


svn path=/trunk/; revision=10480