Initial steps toward adding Plcairo to the PLplot OCaml build system
This patch makes a number of changes to the OCaml portion of the CMake
build system.  A check is added for ocamlfind (an OCaml compilation aid),
checks are made for Cairo and Gtk+ (lablgtk2) OCaml bindings.

If the extcairo device is built and the Cairo OCaml bindings are found
then the Plcairo OCaml module will be built, which includes support for
using the extcairo driver from OCaml.

The check for OCaml Gtk+ bindings is not currently used but it will be
once Plcairo examples are added.


svn path=/trunk/; revision=10488