OCaml's plshade, plshades and plpoly3 all take bool values when appropriate
This is an API change for these three functions!

The rest of the OCaml PLplot API uses OCaml's bool for function parameters
which are PLBOOL in the C API.  A small error in
bindings/ocaml/plplot_core.idl had them mapped to int rather than bool for a
few functions.

svn path=/trunk/; revision=10517