Change pltr1 and pltr2 to be curry-friendly (OCaml)
This commit moves the (x, y) coordinates to the end of the parameter
list for these two functions.  It is now possible to use these functions
in curried form with plset_pltr.

The examples have been updated, and "mypltr" functions from the
examples have been updated to follow this style as well.

svn path=/trunk/; revision=10524