Add plplot function to create a secure temporary file in a standard cross platform way.
The function always returns a file handle (or -1 on failure). 

If a char ** is passed then the filename will also be returned in the pointer. If the file name is not returned then the file will automatically be deleted once the file handle is closed. If the file name is returned, then it is the users responsibility to delete the file and also to free the memory associated with the string containing the file name. Note that the latter use is discouraged as there are some (theoretical) race conditions that can occur in this case, but we may need to use this in plplot because of the way the code is written. 

The temporary file will be created in the directory given by TMPDIR (UNIX) or TEMP (windows) if they are set. Otherwise it will try P_tmpdir (if defined) and finally a default of /tmp (UNIX) or c:\windows\temp (windows). 



svn path=/trunk/; revision=10585