Correct and improve the documentation of plRotationShear.
This resulted in finding a bug in the old shear calculation which is now
replaced by a simpler shear calculation that gives the same results as the
corrected old logic.

This change means that -dev xcairo (and presumably all other drivers that
depend on plRotationShear) gives the same results (aside from small text
size differences) as -dev xwin for all pages of example 28 for the very
first time!


svn path=/trunk/; revision=10690