Move back to the historical PLINT (as opposed to PLFLT) implementation of pointinpolygon from the current PLINT wrapper for plP_pointinpolygon. The current PLFLT implementation of plP_pointinpolygon is replaced by a PLFLT wrapper for pointinpolygon.
I have done these changes because fill_intersection_polygon uses
pointinpolygon, and that turned up a bug in the PLFLT form (a test point
identical to a polygon vertex returned true).  Therefore, I wanted to test
and debug using the exact historical PLINT form (rather than a PLINT wrapper
for the PLFLT form) since that exact version is the one that has received
the most testing in the past.

plP_pointinpolygon is currently only used by the software fallback gradient
"defined region" approach, and small integer truncation errors would not
make much difference in that case.  Furthermore, I have plans to replace
that approach with one based on fill_intersection_polygon so that in the
future it may be possible to simply drop the current plP_pointinpolygon
PLFLT wrapper for pointinpolygon completely.


svn path=/trunk/; revision=10721