Fix scaling for plP_pointinpolygon.  (2.e9 should generate integer overflows according to my analysis, but 1.e9 should be okay.  However, I had to reduce to 1.e8 for some unknown reason (which I am not going to pursue since plP_pointinpolygon will probably be removed soon in any case).
Fix bugs in notintersect.

The result of these changes is the new notpointinpolygon code works (as well
as the old code) for example 25 so use "#define NEW_NOTPOINTINPOLYGON_CODE"
to compile that new code while avoiding compiling the old code (which I will
retain for a while longer for reference).


svn path=/trunk/; revision=10724