Implement PL_NEAR_PARALLEL crossing status.
Implement fill_status to keep track of whether a given split of polygon 2
is known to have filled areas or known to not have filled areas.  At the
final stage where there are not crossings left between the polygons, this
status helps to quickly decide whether to fill or not.

Prepare calls to fill_intersection_polygon so that polygon 1 always
starts outside polygon 2.  That and the positive orientation requirement
considerably simplify the logic of fill_intersection_polygon.

Implement number_crossings to make sure that the crossings are determined
in the correct order for a give segment of polygon 1.

N.B.  These changes build, but still produce incorrect fill results for some
of the sub-pages of example 25.  So this commit is in the nature of a code
drop to keep track of the many changes that have been done with the
-DUSE_FILL_INTERSECTION_POLYGON=ON approach rather than a final result.


svn path=/trunk/; revision=10730