These changes are due to a diagnostic printout of example 6 that correlates the ascii code used for plpoin with the corresponding number used by plsym in example 7.  These changes are made to make plpoin correspond more closely with its stated goal which is that codes between 32 and 127 correspond to the printable ASCII character with that code.
1. Code 92 corresponds to number 729.  Therefore, change 729 in this table
from unicode 0x22C5 (DOT OPERATOR) to 0x005C (REVERSE SOLIDUS = backslash) to
correspond to ascii character 92 which is a backslash.

2. Code 94 corresponds to number 718.  Therefore, change 718 in this table
from unicode 0x00B0 (DEGREE SIGN) to 0x005E (CIRCUMFLEX ACCENT) to
correspond to ascii character 94 which is a circumflex = caret.

Caveats:

A. These fixes contradict what occurs for example 07 for Hershey font
devices (e.g., -dev xwin), but that is an error in our Hershey font drawing
data in the C code in fonts/*.c that is used to generate the Hershey fonts
that should also be fixed to make example 6 come out right for ascii codes
92 and 94.

B. Ascii code 95 should correspond to an underscore. However, nothing can be
done about that via this table because the translation from ascii code to
number is an integral part of the Hershey fonts, and a mistake was made
there such that code 95 corresponds to code 729 which is already used for
backslash.  So this is one other error that must be addressed by fixing a
bug in our Hershey font generation source code.

C. During this investigation it was discovered that for the non-extended
Hershey font table, an additional ~100 codes are used in example 7 that have
no counterpart in this table.  That issue will be addressed in the next
commit to this table.


svn path=/trunk/; revision=10889