Display the union of all the Hershey glyph possibilities for the Hershey font index range from 0 to 2999. Use the standard Hershey font for page 1 (0-99) and the new page 2 (100-199), and the extended Hershey font for the remaining pages (excluding pages where no glyphs exist at all for either font) as before.
Explanation: The standard font and extended font render the same characters
for indices 1-50, but the extended font has no glyphs for indices 51-171
while the standard font does have glyphs for that range.  For indices from
200 to 2999 the reverse is the case; the standard font has no glyphs while
(except for the excluded pages where with special variations of this example
show there are no glyphs for either font) the extended fonts do have glyphs
in that range. Thus, to show the union of all glyphs available for the
standard fonts and extended fonts, the standard font is chosen for the first
two pages and the extended font for the remaining pages.

Please propagate these changes to our standard example 7 to all other
languages.


svn path=/trunk/; revision=10890