Address long-standing issues in ascii index for plpoin to make results conform as closely as possible to the ideal that is documented which is codes between 32 and 127 should correspond to the corresponding ascii character.
92 changed from dot to backslash (804 in the Hershey index)
94 changed from degree symbol to caret (832 in the Hershey index)
95 changed from dot to 590 in the Hershey index.  

N.B. 590 is a placeholder for the underscore.  There is currently no Hershey
glyph at that position, but in future an underscore glyph should be put
there.  Note, 590 is the Hershey index used by PGplot for the underscore
so the 590 placeholder for underscore should assure future compatibility
with PGplot in this regard.

The new results for plstnd5.fnt and plxtnd5.fnt were generated using
-DBUILD_HERSHEY_FONTS=ON and running the resulting executables
./generate_xtndfonts, and ./generate_stndfonts in the fonts subdirectory
of the build tree and copying the results back to the data subdirectory
of the source tree.


svn path=/trunk/; revision=10893