Add some minimal documentation of buffer (subbuffer); findex (subindex); and hersh (subhersh); that are used throughout fonts/font??. to define the Hershey fonts and various indexing schemes for them that are used by Plplot.
Add underscore glyph in Hershey fonts.  Thanks to David MacMahon for
the implementation.

Move underscore position from 590 (which matched PGplot) to 795. This
changes the corresponding shortened index position of the underscore in the
standard fonts from 51 to 133.  This change results in a much more logical
layout of the standard font that is displayed in first two pages of example
7. 795 is also a reasonable position for the underscore in the extended
Hershey font index which is displayed by the remaining pages of example 7
(see especially page 6 of that example).


svn path=/trunk/; revision=10898