Add basic global/universal coordinate transform support
This commit adds a function plstransform which allows a user to define a
global coordinate transform which will be used by all PLplot plotting
functions which act inside the plotting window.  C example 19 has been updated
with an extra page to illustrate the use of plstransform.

This commit also adds the plpath function which acts like pljoin when no
transform is defined and approximates the translated path with a given number
of line segments when a transform is defined.

The OCaml bindings have been updated to include support for plstransform and
plpath.  OCaml example 19 has been updated to match the updated C example.  No
other language bindings are updated by this commit.

The API documentation has been updated to reflect these changes.

svn path=/trunk/; revision=10905