Tweak the wincairo message to remove the WARNING label from it since this is just an informational message concerning a condition (this is not a Windows platform) which the user can do nothing about on that platform.
We should also systematically drop the WARNING label from messages on other
platforms as well if they are just informational. Thus, WARNING messages
should be reserved for something the user has control over such as
downloading/installing more software.  The ideal here is no WARNING messages
at all for fully loaded systems.  With this change we have reestablished
that ideal on Linux, but this needs to be checked for other platforms as
well.


svn path=/trunk/; revision=10912