Implement an example to display all Type 1 glyphs that are potentially accessible to the ps(c) device in a similar manner to the historical gfontview application.  The ps.c file must be edited to #define PL_TEST_TYPE1 in order to make this example work properly.
The point of this example is to duplicate the functionality of the gfontview
application which is no longer being actively developed. Because of this
app's dependence on freetype1 (as opposed to modern freetype) modern distros
have been forced to abandon this useful application which is why I have had
to emulate it with test_type1.py.

If this example (with ps built with PL_TEST_TYPE1 #defined) shows a type1
glyph that would be useful to PLplot, then it is straightforward to edit the
unicode to type1 index translation table at include/plunicode-type1.h to
make that glyph accessible to PLplot by unicode index for ordinary builds of
ps that do not #define PL_TEST_TYPE1.


svn path=/trunk/; revision=10929