Permanantly remove the historical momentary change to symbol font inside plhrsh (used by both plpoin and plsym in examples 6 and 7).  This removal is consistent with what has been done in plcore.c since revision 10623 for PLplot text escape sequences (Hershey, hexadecimal, and Greek).
The original motivation of this momentary change was Type 1 fonts (used for
our ps and pdf device drivers) collect most mathematical symbols in one
"symbol" font, but I have plans to deal with that issue a different way (if
the ordinary Type 1 font lookup fails to get a hit for these device drivers,
then always check the Type 1 symbol font for a hit as a last resort).
Meanwhile, the removal of the built-in momentary change to symbol fonts
allows more consistent font results for our modern devices.

This removal gave good results for examples 6 and 7 for
both modes (pls->alt_unicode = 0 or 1) of unicode character rendering
for -dev pscairo.  Other checks seemed fine as well.

ps and pdf device drivers currently work but with no access to the
mathematical glyphs in the Type1 symbol font.  That issue has existed for
text since revision 10623 and for plpoin and plsym since the current change,
but I plan to address this issue shortly.


svn path=/trunk/; revision=10933