Change the Type1 fonts (used just for the ps and pdf device drivers) corresponding to the "script" and "symbol" FCI font family values to san-serif (Helvetica) to be consistent with the modern device drivers that use TrueType fonts.
A byproduct of this change is the Type1 symbol font is no longer directly
accessible to the PLplot user.  However, ps and pdf device driver changes
are planned to make the Type1 symbol font indirectly available for all FCI
choices for these devices as the fallback when the glyph lookup fails for
the ordinary Type 1 fonts.


svn path=/trunk/; revision=10934