Tweak commentary.
Drop explicit setting of -dev ps -drvopt hrshsym=0.  This allows this
example to display the first 256 unicode glyphs for any unicode-aware
device.  It's only when ps.c is edited to #define PL_TEST_TYPE1 and the
command-line options -dev ps -drvopt hrshsym=0 are used that the unicode
indices are interpreted as Type 1 font indices so that the PostScript
results look like those from the now defunct gfontview application.

Fix a bug where the escape code was not escaped.


svn path=/trunk/; revision=10948