Explicitly specify pltr as "NULL" for plshade and plshades which translates to NULL at the C level and which is interpreted there as defining the x and y arrays from xmin, xmax, nx and ymin, ymax, and ny.  This is very different from using pltr0 (the default user-friendly Tcl result if nothing is specified for pltr).
This solves the Tcl issue of dropping the pltr_data tests for NULL at the C
level.  We were relying on those tests before to obtain the xmin, xmax,...
result but when those tests were dropped we were getting the Tcl
user-friendly pltr0 default instead which had such an incorrect scale and
offset for these examples that the pages were blank or might have a single
polygon.


svn path=/trunk/; revision=10964