Change back to a full GUI setup for all noninteractive qt devices.
There are two reasons for this change.  (1) The non GUI setup for
noninteractive qt devices did not help to solve Hazen's current hang issue
for qt noninteractive devices. (2) Usually a non-GUI setup for the
noninteractive qt devices is fine, but for the special case of multistream use
where the first stream uses a noninteractive qt device and a subsequent
stream uses an interactive (qtwidget) qt device you must have a full GUI
setup for the initial noninteractive qt device.  The reason is the first
stream to use a qt device creates the QApplication object and all others
make use of it.  (Thanks to Andrew Ross for help with this summary).


svn path=/trunk/; revision=11050