Revert previous change which didn't do what we wanted.  The result was
* The PLplot Plotting Library: (plplotdoc).   The PLplot Plotting Library

got changed to

* The PLplot Plotting Library: (plplotdoc).   PLplot Plotting Library

(i.e. where the second "The" is dropped) rather than the desired

* PLplot Plotting Library: (plplotdoc).   The PLplot Plotting Library

(i.e., where the first "The" is dropped).

It looks like to drop the first "The" it would be necessary to change
the overall title which would be bad since that would affect html, PostScript,
etc., results as well as the info results.


svn path=/trunk/; revision=11064