Provide a script to run all Tcl examples in the same way as we do for the Tk examples. The script properly resets PLplot so that each example displays correctly (with the exception of x21). Reformatted the Tk script, as the list of commands to reset PLplot got fairly long.
Currently, the script (standard_examples) can not be used as a proper shell script, in the same way as the 
Tk equivalent. This is a first version to get the basic functionality.

Re x21: this example does something to upset PLplot on my machine, so that the next example complains that
I should call plinit() again. I have commented it out for now.

svn path=/trunk/; revision=11070