Replace xcairo dirty rectangles with optional offscreen (image) buffering
Adds a new drvopt, image_buffering, which is on (= 1) by default.

This also does a bit of minor cleanup in the way drawing properties are
handled.

Disabling image_buffering causes all plot drawing to occur directly on an
X surface (still buffered off-screen).  This should maintain a similar or
better speed increase for example 17 as provided by the dirty rectangle
approach, while avoiding the negative performance impact the dirty
rectangle code had on some routines, such as plimage (example 20).

The cleanup in this patch consists of:

1. Using cairo_save and cairo_restore to save/restore the state of various
   Cairo drawing settings when they are changed within a drawing function.
2. Using the CAIRO_OPERATOR_OVER operator to avoid seams between adjacent
   polygons, thanks to a suggestion on the Cairo mailing list.

svn path=/trunk/; revision=11080