Previous commit by Jerry removed all configurable items from Ada examples source files so that the configure of those source files got turned into a copy. Strip out this configure=copy for the Ada examples source files and adjust the Ada examples build accordingly for the changed directory for the examples source.
N.B. the result passes the "make test_noninteractive" test for the
build tree, and that same test for the installed examples with
both the traditional Makefile+pkg-config approach for building and
testing those examples as well as the more modern CMake-based
approach for building and testing those examples.  So all seems well
for all builds and tests of the Ada bindings and examples with the
new non-configured approach.

N.B. One further known change is planned for the website generation to
adjust for the new location of the Ada source files (in the source
tree rather than configured files in the build tree).


svn path=/trunk/; revision=11099