Ada examples now copied directly from source tree since they are no longer configured.
This concludes the changes that I believe are necessary
due to the new non-configured approach to Ada source files that has
been initiated by Jerry.

N.B. this last change will remain untested until the
next attempt to generate the website, but I believe it
will be fine  since the change is so simple.


svn path=/trunk/; revision=11100