Improve Werner's build system for libharu.
Comment out installation of files that (a) do not belong in the
install tree or (b) are installed in the wrong location in the build
tree from the POV of Unix standards.  These are all files that are not
needed for the basic use of the install tree for this library.  For
example, I currently do not install anything to do with the demos.
If/when that is done, it would be good to install those in an
acceptable install-tree location.  That installation should be
accompanied by implementation of an independent build
system to build and test those demos.

Implement versioning system for library.

Implement library and executable RPATH and SONAME support.

Improved name for static library.

Improved handling of target names for static and shared library forms.

Implement build-system support for Cygwin visibility.

POST_2.1.0 option (default to ON) to handle case where additional files
(currently just hpdf_namedict.c) added to list of files being compiled
into library after the 2.1.0 version of libharu was released.


svn path=/trunk/; revision=11113